Commid3.1

+++ ComercialLunaP/src/main/java/comerciallunapazmino/com/ComercialLunaP/repository/PedidoCabeceraRepository.java
@@ -20,6 +20,6 @@ public interface PedidoCabeceraRepository extends JpaRepository<PedidosCabeceras
 	//List<PedidosCabeceras> findByEstado(@Param("estado") String estado);
 	
 	@Query("SELECT SUM(total) FROM PedidosCabeceras")
-	Float selectTotals();
+	Double selectTotals();
 	
 }
